Write both schemas from same source
